feat: Separate Callback page components and apply Suspense
